refactor: ステージ生成ロジックを分離
